+++ src/simplebot/commands.py
@@ -4,6 +4,7 @@ from collections import OrderedDict
 from typing import Callable, Dict, Generator, Optional, Set
 
 from .hookspec import deltabot_hookimpl
+from .templates import help_template
 
 CMD_PREFIX = "/"
 _cmds: Set[tuple] = set()
@@ -123,16 +124,14 @@ class Commands:
     def command_help(self, bot, command, replies) -> None:
         """reply with help message about available commands."""
         is_admin = bot.is_admin(command.message.get_sender_contact().addr)
-        l = []
-        l.append("➡️ Commands:\n")
+        cmds = []
         for c in self._cmd_defs.values():
             if not c.admin or is_admin:
-                l.append("{}: {}\n".format(c.cmd, c.short))
-        l.append("\n\n")
+                cmds.append(c)
         pm = bot.plugins._pm
         plugins = [pm.get_name(plug) for plug, dist in pm.list_plugin_distinfo()]
-        l.append("🧩 Enabled Plugins:\n{}".format("\n".join(plugins)))
-        replies.add(text="\n".join(l))
+        html = help_template.render(addr=bot.self_contact.addr, cmds=cmds, plugins=plugins)
+        replies.add(text="ℹ️ Help", html=html)
 
 
 class CommandDef: