+++ src/simplebot/templates/__init__.py
@@ -1,0 +1,7 @@
+from jinja2 import Environment, PackageLoader, select_autoescape
+
+env = Environment(
+    loader=PackageLoader(__name__.split(".")[0], "templates"),
+    autoescape=select_autoescape(["html", "xml"]),
+)
+help_template = env.get_template("help.j2")