+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - ignore messages from other bots using the new Delta Chat API.
 - allow to get account configuration values with `set_config` command.
-- improved command and filter registration.
+- **breaking change:** improved command and filter registration.
 - allow to register administrators-only filters.
 
 ## [1.1.1]